requirements.txt
aldryn-search
cmsplugin-iframe2
django-cms-articles==2.1.4
leprikon
